django-angular/backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Nicolas M. <mine.nicolas@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.7"
djangorestframework = "^3.15.2"
django-admin-sortable2 = "^2.2.1"
django-orderable = "^6.1.2"
pandas = "^2.2.2"
django-slick-reporting = "^1.3.1"
django-crispy-forms = "^2.3"
crispy-bootstrap4 = "^2024.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
